第五次提交(exec导入导出)

+++ zbf-common/src/main/java/com/zbf/common/entity/my/BaseMenu.java
@@ -67,6 +67,6 @@ public class BaseMenu implements Serializable {
     private String loginName;
 
     @TableField(exist = false)
-    private Integer bid;
+    private Long bid;
 
 }
